+++ src/modules/ui/checkbox-button/index.tsx
@@ -4,15 +4,13 @@ import { Typography } from '~/modules/ui/typography'
 import { Icon } from '~/modules/ui/icon'
 import { cn } from '~/utils/cn'
 
-export default function CheckboxButton({
-  label,
-  onPress,
-  checked = false,
-}: {
+interface CheckboxButtonProps {
   label: string
   onPress: () => void
   checked?: boolean
-}) {
+}
+
+export default function CheckboxButton({ label, onPress, checked = false }: CheckboxButtonProps) {
   const buttonClassName = cn([
     'rounded-full flex flex-row items-center gap-2 bg-button-secondary-bg min-h-md px-4 py-3',
     checked && 'bg-bg-vibrant-accent',
