+++ src/modules/ui/checkbox-list/index.tsx
@@ -1,16 +1,14 @@
 import { View } from 'react-native'
 import React from 'react'
-import CheckboxButton from '../CheckboxButton'
+import CheckboxButton from '../checkbox-button'
 
-export function CheckboxList({
-  options,
-  selectedOption,
-  onSelect,
-}: {
+interface CheckboxListProps {
   options: string[]
   selectedOption: string | null
   onSelect: (label: string) => void
-}) {
+}
+
+export function CheckboxList({ options, selectedOption, onSelect }: CheckboxListProps) {
   return (
     <View className="gap-4 flex-1">
       {options.map((label) => (
